Alex Ewerlöf @userpixel (github.com/userpixel)
https://codeburst.io/async-map-with-limited-parallelism-2b91bd47af70

https://github.com/userpixel/cap-parallel

Sweden (Swedish: Sverige [ˈsvæ̌rjɛ]), officially the Kingdom of Sweden
(Swedish: Konungariket Sverige [ˈkôːnɵŋaˌriːkɛt ˈsvæ̌rjɛ]), is a Nordic
country in Northern Europe. It borders Norway to the west and north,
Finland to the east, and is connected to Denmark in the southwest by a
bridge-tunnel across the Öresund Strait. At 450,295 square kilometres
(173,860 sq mi), Sweden is the largest country in Northern Europe, the
third-largest country in the European Union, and the fifth largest
country in Europe. The capital and largest city is Stockholm. Sweden
has a total population of 10.4 million; and a low population density of
25 inhabitants per square kilometre (65/sq mi). 87% of Swedes live in
urban areas, which cover 1.5% of the entire land area. The highest
concentration is in the central and southern half of the country.

― Wikipedia

Signed-off-by: Benjamin Vincent (Luxcium) <luxcium@neb401.com>